BTCPAY server Integration

+++ packages/server/src/controllers/user.js
@@ -3,6 +3,7 @@ const bcrypt = require('bcryptjs');
 
 const User = require('../models/user.js');
 const Alias = require('../models/alias.js');
+const Invoice = require('../models/invoice.js')
 
 const emailLib = require('../lib/email.js');
 const errorLib = require('../lib/error.js');
@@ -19,12 +20,39 @@ exports.getAliases = async (req, res, next) => {
 	}
 }
 
-exports.renewAlias = async (req, res, next) => {
-	const alias = req.params.alias;
-	const domain = req.body.domain;
+exports.getInvoices = async (req, res, next) => {
+	try {
+		const userObject = await User.findById(req.user.id)
 
+		const invoices = await Invoice.find({ _id: userObject.invoices }).populate('alias')
+
+		const mappedInvoices = []
+		invoices.map(invoice => {
+			mappedInvoices.push({
+				plan: invoice.plan,
+				state: invoice.state,
+				invoiceId: invoice.invoiceId,
+				alias: `${invoice.alias.alias}.${invoice.alias.domain}`,
+				date: invoice.createdAt
+			})
+		})
+
+		const filteredInvoices = mappedInvoices.filter(invoice => invoice.state == 'InvoiceSettled' || invoice.state == 'InvoiceProcessing');
+
+		res.status(200).json(filteredInvoices);
+	}
+	catch (err) {
+		next(errorLib.errorWrapper(err));
+	}
+}
+
+exports.renewAlias = async (req, res, next) => {
+	const { alias } = req.params;
+	const { domain } = req.query;
 	try {
 		const aliasObject = await Alias.findOne({ alias, domain, user: req.user.id });
+
+		if(aliasObject.paid) throw errorLib.conflictError('You cannot renew a paid alias')
 
 		const sevenDays = 604800000; // In milliseconds
 		const now = new Date();
@@ -80,6 +108,8 @@ exports.deleteAlias = async (req, res, next) => {
 
 	try {
 		const aliasObject = await Alias.findOne({ alias: alias, domain: domain, user: req.user.id });
+
+		if(aliasObject.paid) throw errorLib.conflictError('You cannot delete a paid alias')
 
 		await MongoLib.deleteAlias(aliasObject);
 
@@ -160,7 +190,7 @@ exports.retrieveTwoFactorAuthSecret = async (req, res, next) => {
 		if (user.requireTwoFactor) throw errorLib.unauthorizedAccessError('You cannot view the secret for 2FA');
 
 		const secret = authenticator.generateSecret();
-		const otpauthurl = authenticator.keyuri(user.email,'Cryptachi',secret);
+		const otpauthurl = authenticator.keyuri(user.email, 'Cryptachi', secret);
 
 		user.twoFactorSecret = secret;
 		await user.save();
@@ -177,14 +207,14 @@ exports.disableTwoFactorAuth = async (req, res, next) => {
 		const user = await User.findById(req.user.id);
 
 		if (!user.requireTwoFactor) throw errorLib.unauthorizedAccessError("2FA is already disabled!")
-		
-		const verified = authenticator.check(authCode,user.twoFactorSecret);
-		if(verified){
+
+		const verified = authenticator.check(authCode, user.twoFactorSecret);
+		if (verified) {
 			user.requireTwoFactor = false;
 			user.twoFactorSecret = null;
 			await user.save();
 		}
-		else{
+		else {
 			throw errorLib.authenticationError('You did not enter the correct Auth code')
 		}
 
@@ -195,23 +225,22 @@ exports.disableTwoFactorAuth = async (req, res, next) => {
 }
 
 exports.changePassword = async (req, res, next) => {
-  const password = req.body.password;
+	const password = req.body.password;
 
-  try {
-    const user = await User.findById(req.user.id);
+	try {
+		const user = await User.findById(req.user.id);
 
 		const salt = await bcrypt.genSalt(10);
 		user.password = await bcrypt.hash(password, salt);
 
-    await user.save();
+		await user.save();
 
-    // No need to await.
-    emailLib.sendPasswordChangeNotification(user.email);
+		emailLib.sendPasswordChangeNotification(user.email);
 
 		res.status(200).json({ message: 'Password changed successfully.' });
-  } catch(err) {
-    next(errorLib.errorWrapper(err));
-  }
+	} catch (err) {
+		next(errorLib.errorWrapper(err));
+	}
 }
 
 //user confirms their 2fa by sending the current auth code on their phone
@@ -224,7 +253,7 @@ exports.enableTwoFactorAuth = async (req, res, next) => {
 		if (user.requireTwoFactor) throw errorLib.unauthorizedAccessError("You've already enabled 2FA")
 		if (!user.twoFactorSecret) throw errorLib.unauthorizedAccessError("You cannot activate 2FA at this time")
 
-		const verified = authenticator.check(authCode,user.twoFactorSecret);
+		const verified = authenticator.check(authCode, user.twoFactorSecret);
 
 		if (verified) {
 			user.requireTwoFactor = true;
